// src/styles/types.ts
export type SizesType = {
  XXS: number;
  XS: number;
  S: number;
  M: number;
  L: number;
  XL: number;
  XXL: number;
  XXXL: number;
};

export type LineHeightType = {
  XXS: number;
  XS: number;
  S: number;
  M: number;
  L: number;
  XL: number;
  XXL: number;
  XXXL: number;
};

export type FontFamilyType = {
  REGULAR: string;
  BOLD: string;
  SEMIBOLD: string;
  MEDIUM: string;
  ITALIC: string;
  LIGHT: string;
};